ok, back to doubles. Let's go! That was fun! A quick find/replace. Tests green. Acne gone. Sometimes, I love my hobby...and yes, having tests is amazing! Having them, because you did test first is even more amazing!

+++ src/main/java/org/schakalacka/java/raytracing/geometry/objects/Cube.java
@@ -20,7 +20,7 @@ public class Cube extends Shape {
         if (tmin > tmax) {
             return List.of();
         } else {
-            return List.of(new Intersection(this, (float) tmin), new Intersection(this, (float) tmax));
+            return List.of(new Intersection(this,  tmin), new Intersection(this,  tmax));
         }
     }
