+++ src/main/java/org/schakalacka/java/raytracing/math/NaiveMatrixProvider.java
@@ -5,7 +5,7 @@ import org.schakalacka.java.raytracing.Counter;
 class NaiveMatrixProvider implements IMatrixProvider {
 
 
-    public NaiveMatrix get(float[][] ref) {
+    public NaiveMatrix get(double[][] ref) {
 
         NaiveMatrix result = this.get(ref.length);
 
@@ -35,7 +35,7 @@ class NaiveMatrixProvider implements IMatrixProvider {
     /***
      * create a translation matrix. It is a 4x4 identity matrix, where the last colum is populated with the 3 values.
      */
-    public NaiveMatrix translation(float x, float y, float z) {
+    public NaiveMatrix translation(double x, double y, double z) {
         Counter.translate++;
         NaiveMatrix result = this.get(4, true);
         result.set(0, 3, x);
@@ -49,7 +49,7 @@ class NaiveMatrixProvider implements IMatrixProvider {
 //                {0, 0, 0, 1},
     }
 
-    public NaiveMatrix scaling(float x, float y, float z) {
+    public NaiveMatrix scaling(double x, double y, double z) {
         Counter.scale++;
         NaiveMatrix result = this.get(4, false);
         result.set(0, 0, x);
@@ -68,13 +68,13 @@ class NaiveMatrixProvider implements IMatrixProvider {
      *
      * @return a left-handed rotation matrix along the X axis
      */
-    public NaiveMatrix rotationX(float radians) {
+    public NaiveMatrix rotationX(double radians) {
         Counter.rotX++;
         NaiveMatrix result = this.get(4, true);
-        result.set(1, 1, (float) Math.cos(radians));
-        result.set(1, 2, (float) -Math.sin(radians));
-        result.set(2, 1, (float) Math.sin(radians));
-        result.set(2, 2, (float) Math.cos(radians));
+        result.set(1, 1,  Math.cos(radians));
+        result.set(1, 2,  -Math.sin(radians));
+        result.set(2, 1,  Math.sin(radians));
+        result.set(2, 2,  Math.cos(radians));
 
         return result;
 
@@ -88,20 +88,20 @@ class NaiveMatrixProvider implements IMatrixProvider {
      *
      * @return a left-handed rotation matrix along the Y axis
      */
-    public NaiveMatrix rotationY(float radians) {
+    public NaiveMatrix rotationY(double radians) {
         Counter.rotY++;
         NaiveMatrix result = this.get(4, true);
-        result.set(0, 0, (float) Math.cos(radians));
+        result.set(0, 0,  Math.cos(radians));
         result.set(0, 1, 0);
-        result.set(0, 2, (float) Math.sin(radians));
+        result.set(0, 2,  Math.sin(radians));
         result.set(0, 3, 0);
         result.set(1, 0, 0);
         result.set(1, 1, 1);
         result.set(1, 2, 0);
         result.set(1, 3, 0);
-        result.set(2, 0, (float) -Math.sin(radians));
+        result.set(2, 0,  -Math.sin(radians));
         result.set(2, 1, 0);
-        result.set(2, 2, (float) Math.cos(radians));
+        result.set(2, 2,  Math.cos(radians));
         result.set(2, 3, 0);
         result.set(3, 0, 0);
         result.set(3, 1, 0);
@@ -116,13 +116,13 @@ class NaiveMatrixProvider implements IMatrixProvider {
 //                {0, 0, 0, 1},
     }
 
-    public NaiveMatrix rotationZ(float radians) {
+    public NaiveMatrix rotationZ(double radians) {
         Counter.rotZ++;
         NaiveMatrix result = this.get(4, true);
-        result.set(0, 0, (float) Math.cos(radians));
-        result.set(0, 1, (float) -Math.sin(radians));
-        result.set(1, 0, (float) Math.sin(radians));
-        result.set(1, 1, (float) Math.cos(radians));
+        result.set(0, 0,  Math.cos(radians));
+        result.set(0, 1,  -Math.sin(radians));
+        result.set(1, 0,  Math.sin(radians));
+        result.set(1, 1,  Math.cos(radians));
 
         return result;
 
@@ -133,7 +133,7 @@ class NaiveMatrixProvider implements IMatrixProvider {
 //                {0, 0, 0, 1},
     }
 
-    public NaiveMatrix shearing(float xy, float xz, float yx, float yz, float zx, float zy) {
+    public NaiveMatrix shearing(double xy, double xz, double yx, double yz, double zx, double zy) {
         Counter.shear++;
         NaiveMatrix result = this.get(4, true);
         result.set(0, 1, xy);
