+++ src/main/java/org/schakalacka/java/raytracing/math/Tuple.java
@@ -7,16 +7,16 @@ import static org.schakalacka.java.raytracing.Constants.EQUALS_EPSILON;
 // TODO refactor this into an interface and make point and vector first class citizens
 public class Tuple {
 
-    private final float[] values = new float[4];
+    private final double[] values = new double[4];
 
-    public Tuple(final float x, final float y, final float z, final float w) {
+    public Tuple(final double x, final double y, final double z, final double w) {
         values[0] = x;
         values[1] = y;
         values[2] = z;
         values[3] = w;
     }
 
-    public static Tuple tuple(float x, float y, float z, float w) {
+    public static Tuple tuple(double x, double y, double z, double w) {
         if (w == 0) {
             return new RTVector(x, y, z);
         } else if (w == 1) {
@@ -26,27 +26,27 @@ public class Tuple {
         }
     }
 
-    public static RTPoint point(float x, float y, float z) {
+    public static RTPoint point(double x, double y, double z) {
         return new RTPoint(x, y, z);
     }
 
-    public static RTVector vector(float x, float y, float z) {
+    public static RTVector vector(double x, double y, double z) {
         return new RTVector(x, y, z);
     }
 
-    public float x() {
+    public double x() {
         return values[0];
     }
 
-    public float y() {
+    public double y() {
         return values[1];
     }
 
-    public float z() {
+    public double z() {
         return values[2];
     }
 
-    public float w() {
+    public double w() {
         return values[3];
     }
 
@@ -70,24 +70,24 @@ public class Tuple {
         return Tuple.tuple(this.x() * -1, this.y() * -1, this.z() * -1, this.w() * -1);
     }
 
-    public Tuple mul(float scalar) {
+    public Tuple mul(double scalar) {
         return Tuple.tuple(this.x() * scalar, this.y() * scalar, this.z() * scalar, this.w() * scalar);
     }
 
-    public Tuple div(float scalar) {
+    public Tuple div(double scalar) {
         return Tuple.tuple(this.x() / scalar, this.y() / scalar, this.z() / scalar, this.w() / scalar);
     }
 
-    public float magnitude() {
-        return (float) Math.sqrt(Math.pow(x(), 2) + Math.pow(y(), 2) + Math.pow(z(), 2) + Math.pow(w(), 2));
+    public double magnitude() {
+        return  Math.sqrt(Math.pow(x(), 2) + Math.pow(y(), 2) + Math.pow(z(), 2) + Math.pow(w(), 2));
     }
 
     public Tuple normalize() {
-        final float magnitude = magnitude();
+        final double magnitude = magnitude();
         return Tuple.tuple(this.x() / magnitude, this.y() / magnitude, this.z() / magnitude, this.w() / magnitude);
     }
 
-    public float dot(Tuple that) {
+    public double dot(Tuple that) {
         return this.x() * that.x()
                 + this.y() * that.y()
                 + this.z() * that.z()
@@ -138,7 +138,7 @@ public class Tuple {
 
     }
 
-    public float get(int i) {
+    public double get(int i) {
         return values[i];
     }
 
@@ -146,7 +146,7 @@ public class Tuple {
         return this.sub(that.mul(2).mul(this.dot(that)));
     }
 
-    public float[] getArray() {
+    public double[] getArray() {
         return this.values;
     }
 }
